+++ R/crossSpatialInference.R
@@ -51,6 +51,7 @@
 #'                       rSeq = seq(0, 50, length.out = 50), correction = "rs",
 #'                       sample_id = "patient_id",
 #'                       image_id = "image_number", condition = "patient_stage",
+#'                       algorithm = "bam",
 #'                       ncores = 1
 #'                   )
 #'