+++ R/plotMdl.R
@@ -55,7 +55,8 @@
 #'     data = dat, x = metricRes$r |> unique(),
 #'     designmat = designmat, weights = dat$npoints,
 #'     formula = formula(Y ~ conditionLong_duration +
-#'         conditionOnset + s(patient_id, bs = "re"))
+#'         conditionOnset + s(patient_id, bs = "re")),
+#'         algorithm = "bam"
 #' )
 #' summary(mdl)
 #' plotLs <- lapply(colnames(designmat), plotMdl,